pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "amplifier-module-tool-nano-banana"
version = "0.1.0"
description = "Nano Banana Pro VLM tool for Amplifier - mockup analysis and visual comparison"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Ken", email = "ken@example.com" }
]
keywords = ["amplifier", "ai", "vlm", "gemini", "mockup", "design"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "pyright>=1.1.0",
]

[project.urls]
Homepage = "https://github.com/microsoft/amplifier-module-tool-nano-banana"
Repository = "https://github.com/microsoft/amplifier-module-tool-nano-banana"
Issues = "https://github.com/microsoft/amplifier-module-tool-nano-banana/issues"

[tool.hatch.build.targets.wheel]
packages = ["amplifier_module_tool_nano_banana"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = []

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.11"
